requirements.txt
requests>=2.31
duckdb>=1.1
python-dotenv>=1.0
